Allow affiliate users signup
