+++ backend/db/migrate/20190827141052_add_social_media_url_to_users.rb
@@ -1,0 +1,5 @@
+class AddSocialMediaUrlToUsers < ActiveRecord::Migration[5.1]
+  def change
+    add_column :users, :social_media_url, :string
+  end
+end
